feat(onboarding): add pronouns, birthday, and privacy toggles to make-it-yours step
Extracts InlineBirthday from SettingsScreen into a shared component so the
onboarding profile step and settings screen both use the same edit UI.
